+++ aws-java-sdk-code-generator/src/main/java/com/amazonaws/codegen/model/service/ServiceMetadata.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016. Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
 
 package com.amazonaws.codegen.model.service;
 
+import java.util.List;
 import java.util.Map;
 
 public class ServiceMetadata {
@@ -34,6 +35,8 @@ public class ServiceMetadata {
     private String xmlNamespace;
 
     private String protocol;
+
+    private List<String> protocols;
 
     private String jsonVersion;
 
@@ -107,12 +110,28 @@ public class ServiceMetadata {
         this.xmlNamespace = xmlNamespace;
     }
 
+    /**
+     * {@code protocol} superseded by {@code protocols} field, resolved in {@code ProtocolSelector}
+     */
+    @Deprecated
     public String getProtocol() {
         return protocol;
     }
 
+    /**
+     * {@code protocol} superseded by {@code protocols} field, resolved in {@code ProtocolSelector}
+     */
+    @Deprecated
     public void setProtocol(String protocol) {
         this.protocol = protocol;
+    }
+
+    public List<String> getProtocols() {
+        return protocols;
+    }
+
+    public void setProtocols(List<String> protocols) {
+        this.protocols = protocols;
     }
 
     public String getJsonVersion() {