+++ aws-java-sdk-code-generator/src/main/java/com/amazonaws/codegen/naming/DefaultNamingStrategy.java
@@ -25,10 +25,10 @@ import static com.amazonaws.codegen.internal.Utils.unCapitialize;
 
 import com.amazonaws.codegen.internal.Utils;
 import com.amazonaws.codegen.model.config.customization.CustomizationConfig;
-import com.amazonaws.codegen.model.intermediate.Protocol;
 import com.amazonaws.codegen.model.service.Input;
 import com.amazonaws.codegen.model.service.Output;
 import com.amazonaws.codegen.model.service.ServiceModel;
+import com.amazonaws.codegen.utils.ProtocolUtils;
 import com.amazonaws.util.StringUtils;
 import java.util.Arrays;
 import java.util.Collections;
@@ -152,7 +152,8 @@ public class DefaultNamingStrategy implements NamingStrategy {
 
     @Override
     public String getFluentSetterMethodName(String memberName) {
-        if (Protocol.fromValue(serviceModel.getMetadata().getProtocol()) == Protocol.API_GATEWAY) {
+        String protocolValue = ProtocolUtils.resolveProtocol(serviceModel.getMetadata());
+        if ("api-gateway".equals(protocolValue)) {
             return Utils.unCapitialize(memberName);
         } else {
             return String.format("with%s", Utils.capitialize(memberName));